requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
rich==13.7.0
cryptography==42.0.5
pyyaml==6.0.1
pyjwt>=2.3.0
tldextract>=3.4.4
dnspython==2.6.1
aiohttp_socks>=0.8.4
numpy>=1.26.2
scikit-learn>=1.3.2
faker==24.1.0
playwright==1.42.0
requests==2.31.0
asyncio>=3.4.3
urllib3==2.2.1
bs4>=0.0.2
charset-normalizer>=3.4.0
python-dateutil>=2.8.2
colorama==0.4.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-magic-bin>=0.4.14
fernet==1.0.1
pyOpenSSL==24.0.0
tqdm==4.66.2
python-dotenv==1.0.1
